refactor: extract insights net worth card
